+++ frontend-osori/src/components/layout/BottomNavigationBar.vue
